// projet/src/app/core/model/villageois.ts
/**
 * 24h2025
 * Contrat API des 24h du code 2025.   Version Participant.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TypeVillageois } from './typeVillageois';
import {ActionResponse} from '../../shared/services/api-backend.service';

export interface Villageois {
    /**
     * Identifiant technique du villageois
     */
    idVillageois: string;
    /**
     * Nom du villageois.
     */
    nom: string;
    /**
     * Date de la dernière action du villageois
     */
    dateDerniereAction: string;
    /**
     * Indique si le villageois peut réaliser une action
     */
    disponible: boolean;
    type: TypeVillageois;
    /**
     * Position X du villagois
     */
    positionX?: number;
    /**
     * Position Y du villagois
     */
    positionY?: number;

    currentAction?: ActionResponse;
}